requirements.txt
pyBarcode==0.7
Pillow==3.4.2
